Change ArtistRankDataSource to ignore initial parameter
This is just a hint and it was requesting the wrong values on refresh.
I've changed it to use the current date for refresh and it works great.
Only downside is most of the tests are now migrated to use loadAfter
instead of loadInitial since there's no way to specify the parameter now