Add view data class for play history network state
This is closer to what I'd like for view models. I'm not entirely
happy with two states for both top and bottom because it adds code
duplication. But I wouldn't really want just a flag on the two
states because I think it'll make when statements more complicated
and hide the difference