Add shimmer to display loading on profile
The UI is getting pretty cluttered now. It annoys me that I have
to duplicate the layout for the shimmer but this seems to be the
recommended way and also a good way to get the tests to work well too.
I'll move things into separate files and use includes soon.