Change feed to be able to open artists
The tap interaction should be sent to the viewmodel to decide whether
to open the artist or not, but this was a little complicated.
I couldn't come up with a good way for the view to pass back the
index to the view model and I also didn't like the view model reading
into the paged list. It just seems like a lot of communication overhead
when this functionality's unlikely to change.

I feel bad that there's no new tests for this, but I'll leave this is
until it becomes a problem